+++ gaphor/SysML/__init__.py
@@ -1,3 +1,4 @@
+import gaphor.SysML.deletable
 import gaphor.SysML.diagramlabel
 import gaphor.SysML.drop
 import gaphor.SysML.iconname