+++ gaphor/SysML/diagramitems.py
@@ -1,3 +1,4 @@
 from gaphor.SysML.allocations import *
 from gaphor.SysML.blocks import *
 from gaphor.SysML.requirements import *
+from gaphor.SysML.diagramframe import DiagramFrameItem